refactor: consolidate byte-identical duplicated helpers into shared modules
Moves verified byte-identical functions (diffed/hashed before merging) out
of per-provider and per-endpoint files into shared modules:
_confirmed_keys, _pair_scope/_safe_scope/_is_capture_mode, chunk helper,
_pick_instance_id/_merge_instance_block, and ISO/watermark helpers into
providers/sync/_mod_common.py; sink helpers into new
providers/scrobble/_sink_common.py; webhook helpers into
providers/webhooks/_utils.py (cache dicts kept per-module, only get/put
logic shared); _tmdb_api_key into cw_platform/config_base.py; twoway
orchestrator helpers now import from oneway instead of redefining;
_applier.py's apply_add/apply_update/apply_remove unified into one
parameterized _apply_op.

Copies with real behavioral drift (crosswatch/anilist scope handling,
mdblist chunk semantics, per-provider _emit/_as_bool differences, etc.)
were deliberately left unconsolidated. No behavior change; full suite
matches pre-existing baseline (231 passed, 3 known-unrelated failures),
all touched modules import cleanly.
