+++ providers/scrobble/mdblist/sink.py
@@ -3,7 +3,7 @@
 # Copyright (c) 2025-2026 CrossWatch / Cenodude
 from __future__ import annotations
 
-from collections.abc import Callable, Mapping
+from collections.abc import Callable
 import json
 import time
 from pathlib import Path
@@ -11,7 +11,6 @@ from typing import Any
 
 import requests
 
-from cw_platform.config_base import load_config
 from cw_platform.provider_instances import normalize_instance_id
 from services.activity import record_scrobble_event
 
@@ -21,6 +20,15 @@ except Exception:
     BASE_LOG = None
 
 from providers.scrobble._auto_remove_watchlist import remove_across_providers_by_ids as _rm_across
+from providers.scrobble._sink_common import (
+    _app_meta,
+    _cfg,
+    _cfg_delete_enabled,
+    _clamp,
+    _extract_skeleton_from_body,
+    _is_debug,
+    _merged_provider_block,
+)
 try:
     from providers.scrobble.scrobble import ScrobbleSink, ScrobbleEvent, mask_account as _mask_account  # type: ignore
 except ImportError:
@@ -46,24 +54,10 @@ _AR_TTL = 60
 _TVDB_SHOW_ID_MAX = 9_999_999
 
 
-def _cfg() -> dict[str, Any]:
-    try:
-        return load_config()
-    except Exception:
-        return {}
-
-
 def _provider_auth():
     from providers.auth import runtime as provider_auth
 
     return provider_auth
-
-
-def _is_debug() -> bool:
-    try:
-        return bool((_cfg().get("runtime") or {}).get("debug"))
-    except Exception:
-        return False
 
 
 def _log(msg: str, lvl: str = "INFO") -> None:
@@ -77,29 +71,6 @@ def _log(msg: str, lvl: str = "INFO") -> None:
         except Exception:
             pass
     print(f"[MDBLIST:{level}] {msg}")
-
-
-def _merged_provider_block(cfg: Mapping[str, Any], key: str, instance_id: Any = None) -> dict[str, Any]:
-    base = cfg.get(key) if isinstance(cfg, Mapping) else None
-    blk = dict(base or {}) if isinstance(base, Mapping) else {}
-    inst = normalize_instance_id(instance_id)
-    if inst != "default":
-        insts = blk.get("instances")
-        if isinstance(insts, Mapping) and isinstance(insts.get(inst), Mapping):
-            overlay = dict(insts.get(inst) or {})
-            blk.pop("instances", None)
-            out = dict(blk)
-            out.update(overlay)
-            return out
-    blk.pop("instances", None)
-    return blk
-
-
-def _app_meta(cfg: dict[str, Any]) -> dict[str, str]:
-    rt = cfg.get("runtime") or {}
-    av = str(rt.get("version") or APP_AGENT)
-    ad = (rt.get("build_date") or "").strip()
-    return {"app_version": av, **({"app_date": ad} if ad else {})}
 
 
 def _timeout(cfg: dict[str, Any]) -> float:
@@ -175,14 +146,6 @@ def _progress_step(cfg: dict[str, Any]) -> int:
     return max(1, min(25, step_i))
 
 
-def _clamp(p: Any) -> int:
-    try:
-        v = int(float(p))
-    except Exception:
-        v = 0
-    return max(0, min(100, v))
-
-
 def _state_dir() -> Path:
     base = Path("/config/.cw_state") if Path("/config/config.json").exists() else Path(".cw_state")
     try:
@@ -226,28 +189,6 @@ def _norm_type(s: str) -> str:
     if s == "series":
         s = "show"
     return s
-
-
-def _cfg_delete_enabled(cfg: dict[str, Any], media_type: str) -> bool:
-    s = cfg.get("scrobble") or {}
-    watch = s.get("watch") or {}
-    route_opts_raw = watch.get("route_options")
-    route_opts: dict[str, Any] = route_opts_raw if isinstance(route_opts_raw, dict) else {}
-    route_mode = str(route_opts.get("auto_remove_watchlist") or "inherit").strip().lower()
-    if route_mode == "off":
-        return False
-    if not s.get("delete_plex"):
-        if route_mode != "on":
-            return False
-    types = s.get("delete_plex_types") or []
-    mt = _norm_type(media_type)
-    if isinstance(types, str):
-        return _norm_type(types) == mt
-    try:
-        allowed = {_norm_type(x) for x in types if str(x).strip()}
-    except Exception:
-        return False
-    return mt in allowed
 
 
 def _ids(ev: Any) -> dict[str, Any]:
@@ -437,14 +378,6 @@ def _ids_desc_map(ids: dict[str, Any], order: tuple[str, ...]) -> str:
         if v is not None and v != "" and v != 0:
             return f"{k}:{v}"
     return "title/year"
-
-
-def _extract_skeleton_from_body(b: dict[str, Any]) -> dict[str, Any]:
-    out = dict(b)
-    out.pop("progress", None)
-    out.pop("app_version", None)
-    out.pop("app_date", None)
-    return out
 
 
 def _body_ids_desc(b: dict[str, Any]) -> str:
