+++ providers/scrobble/trakt/sink.py
@@ -5,13 +5,22 @@ from __future__ import annotations
 
 from collections.abc import Callable, Mapping
 import json, time
-from pathlib import Path
 from typing import Any
 
 import requests
 
-from cw_platform.config_base import load_config, save_config
+from cw_platform.config_base import save_config
 from cw_platform.provider_instances import normalize_instance_id
+from providers.scrobble._sink_common import (
+    _ar_seen,
+    _ar_state_file,
+    _cfg,
+    _cfg_delete_enabled,
+    _extract_skeleton_from_body,
+    _is_debug,
+    _merged_provider_block,
+    _norm_type,
+)
 
 try:
     from _logging import log as BASE_LOG
@@ -35,18 +44,10 @@ except ImportError:
 TRAKT_API = "https://api.trakt.tv"
 APP_AGENT = "CrossWatch/Watcher/1.0"
 _TOKEN_OVERRIDE: dict[str, str] = {}
-_AR_TTL = 60
 _SENSITIVE_KEYS = {
     "access_token", "refresh_token", "token", "authorization",
     "client_secret", "password", "code", "api_key",
 }
-
-
-def _cfg() -> dict[str, Any]:
-    try:
-        return load_config()
-    except Exception:
-        return {}
 
 
 def _save_cfg(cfg: dict[str, Any]) -> None:
@@ -54,13 +55,6 @@ def _save_cfg(cfg: dict[str, Any]) -> None:
         save_config(cfg)
     except Exception:
         pass
-
-
-def _is_debug() -> bool:
-    try:
-        return bool((_cfg().get("runtime") or {}).get("debug"))
-    except Exception:
-        return False
 
 
 def _log(msg: str, level: str = "INFO") -> None:
@@ -108,22 +102,6 @@ def _safe_log_repr(value: Any) -> str:
         return repr(_redact_log_value(value))
     except Exception:
         return "<unavailable>"
-
-
-def _merged_provider_block(cfg: Mapping[str, Any], key: str, instance_id: Any = None) -> dict[str, Any]:
-    base = cfg.get(key) if isinstance(cfg, Mapping) else None
-    blk = dict(base or {}) if isinstance(base, Mapping) else {}
-    inst = normalize_instance_id(instance_id)
-    if inst != "default":
-        insts = blk.get("instances")
-        if isinstance(insts, Mapping) and isinstance(insts.get(inst), Mapping):
-            overlay = dict(insts.get(inst) or {})
-            blk.pop("instances", None)
-            out = dict(blk)
-            out.update(overlay)
-            return out
-    blk.pop("instances", None)
-    return blk
 
 
 def _app_meta(cfg: dict[str, Any]) -> dict[str, str]:
@@ -312,40 +290,6 @@ def _guid_search(ev: ScrobbleEvent, cfg: dict[str, Any], instance_id: Any = None
     return None
 
 
-def _ar_state_file() -> Path:
-    base = Path("/config/.cw_state") if Path("/config/config.json").exists() else Path(".cw_state")
-    try:
-        base.mkdir(parents=True, exist_ok=True)
-    except Exception:
-        pass
-    return base / "auto_remove_seen.json"
-
-
-def _ar_seen(key: str) -> bool:
-    p = _ar_state_file()
-    try:
-        data = json.loads(p.read_text(encoding="utf-8")) or {}
-    except Exception:
-        data = {}
-    now = time.time()
-    try:
-        data = {k: v for k, v in data.items() if (now - float(v)) < _AR_TTL}
-    except Exception:
-        data = {}
-    if key in data:
-        try:
-            p.write_text(json.dumps(data), encoding="utf-8")
-        except Exception:
-            pass
-        return True
-    data[key] = now
-    try:
-        p.write_text(json.dumps(data), encoding="utf-8")
-    except Exception:
-        pass
-    return False
-
-
 def _ar_key(ids: dict[str, Any], media_type: str, scope: str = "") -> str:
     for k in ("tmdb", "imdb", "tvdb", "trakt", "simkl"):
         v = ids.get(k)
@@ -357,37 +301,6 @@ def _ar_key(ids: dict[str, Any], media_type: str, scope: str = "") -> str:
     except Exception:
         base = f"{media_type}:title/year"
         return f"{scope}|{base}" if scope else base
-
-
-def _norm_type(t: str) -> str:
-    s = (t or "").strip().lower()
-    if s.endswith("s"):
-        s = s[:-1]
-    if s == "series":
-        s = "show"
-    return s
-
-
-def _cfg_delete_enabled(cfg: dict[str, Any], media_type: str) -> bool:
-    s = cfg.get("scrobble") or {}
-    watch = s.get("watch") or {}
-    route_opts_raw = watch.get("route_options")
-    route_opts: dict[str, Any] = route_opts_raw if isinstance(route_opts_raw, dict) else {}
-    route_mode = str(route_opts.get("auto_remove_watchlist") or "inherit").strip().lower()
-    if route_mode == "off":
-        return False
-    if not s.get("delete_plex"):
-        if route_mode != "on":
-            return False
-    types = s.get("delete_plex_types") or []
-    mt = _norm_type(media_type)
-    if isinstance(types, str):
-        return _norm_type(types) == mt
-    try:
-        allowed = {_norm_type(x) for x in types if str(x).strip()}
-    except Exception:
-        return False
-    return mt in allowed
 
 
 def _auto_remove_across(ev: ScrobbleEvent, cfg: dict[str, Any], scope: str = "") -> None:
@@ -454,14 +367,6 @@ def _route_source(cfg: dict[str, Any]) -> tuple[str, str]:
     source = str(watch.get("route_provider") or watch.get("provider") or "watcher").strip().lower() or "watcher"
     source_instance = str(watch.get("route_provider_instance") or watch.get("provider_instance") or "default").strip() or "default"
     return source, source_instance
-
-
-def _extract_skeleton_from_body(b: dict[str, Any]) -> dict[str, Any]:
-    out = dict(b)
-    out.pop("progress", None)
-    out.pop("app_version", None)
-    out.pop("app_date", None)
-    return out
 
 
 def _body_ids_desc(b: dict[str, Any]) -> str:
