+++ providers/sync/_mod_SIMKL.py
@@ -22,6 +22,7 @@ from ._mod_common import (
     parse_rate_limit,
     SimpleRateLimiter,
     request_with_retries,
+    _confirmed_keys,
 )
 from .simkl._common import (
     _pair_scope as simkl_pair_scope,
@@ -34,46 +35,6 @@ from .simkl._common import (
     state_file,
 )
 
-
-def _confirmed_keys(key_of, items: Iterable[Mapping[str, Any]], unresolved: Any) -> list[str]:
-    attempted: list[str] = []
-    for it in items or []:
-        try:
-            k = str(key_of(it) or "").strip()
-        except Exception:
-            k = ""
-        if k:
-            attempted.append(k)
-
-    unresolved_keys: set[str] = set()
-    if unresolved:
-        for u in unresolved:
-            obj: Any = u
-            if isinstance(u, Mapping):
-                if isinstance(u.get("key"), str) and u.get("key"):
-                    unresolved_keys.add(str(u.get("key")))
-                    continue
-                if "item" in u:
-                    obj = u.get("item")
-            if isinstance(obj, str) and obj:
-                unresolved_keys.add(obj)
-                continue
-            if isinstance(obj, Mapping):
-                try:
-                    k = str(key_of(obj) or "").strip()
-                except Exception:
-                    k = ""
-                if k:
-                    unresolved_keys.add(k)
-
-    out: list[str] = []
-    seen: set[str] = set()
-    for k in attempted:
-        if k in unresolved_keys or k in seen:
-            continue
-        out.append(k)
-        seen.add(k)
-    return out
 
 __VERSION__ = "1.1"
 __all__ = ["get_manifest", "SIMKLModule", "OPS"]
