+++ providers/sync/_mod_TRAKT.py
@@ -17,46 +17,6 @@ from .trakt._common import (
 from ._log import log as cw_log
 
 
-def _confirmed_keys(key_of, items: Iterable[Mapping[str, Any]], unresolved: Any) -> list[str]:
-    attempted: list[str] = []
-    for it in items or []:
-        try:
-            k = str(key_of(it) or "").strip()
-        except Exception:
-            k = ""
-        if k:
-            attempted.append(k)
-
-    unresolved_keys: set[str] = set()
-    if unresolved:
-        for u in unresolved:
-            obj: Any = u
-            if isinstance(u, Mapping):
-                if isinstance(u.get("key"), str) and u.get("key"):
-                    unresolved_keys.add(str(u.get("key")))
-                    continue
-                if "item" in u:
-                    obj = u.get("item")
-            if isinstance(obj, str) and obj:
-                unresolved_keys.add(obj)
-                continue
-            if isinstance(obj, Mapping):
-                try:
-                    k = str(key_of(obj) or "").strip()
-                except Exception:
-                    k = ""
-                if k:
-                    unresolved_keys.add(k)
-
-    out: list[str] = []
-    seen: set[str] = set()
-    for k in attempted:
-        if k in unresolved_keys or k in seen:
-            continue
-        out.append(k)
-        seen.add(k)
-    return out
-
 try:
     from ..auth._auth_TRAKT import PROVIDER as AUTH_TRAKT  # token refresh hook
 except Exception:
@@ -82,6 +42,7 @@ from ._mod_common import (
     label_trakt,
     SimpleRateLimiter,
     make_snapshot_progress,
+    _confirmed_keys,
 )
 
 try:  # type: ignore[name-defined]
