+++ providers/sync/trakt/_watchlist.py
@@ -18,11 +18,10 @@ from ._common import (
     _pair_scope,
     _is_capture_mode,
     _now_iso,
-    _chunk,
     _record_limit_error,
     headers_for_adapter,
 )
-from .._mod_common import request_with_retries
+from .._mod_common import request_with_retries, _chunk_items as _chunk
 from cw_platform.id_map import minimal as id_minimal
 from .._log import log as cw_log
