+++ providers/webhooks/plextrakt.py
@@ -31,6 +31,20 @@ try:
 except Exception:
     _rm_across_api = None
 
+from providers.webhooks._utils import (
+    _app_meta,
+    _best_id_key_order,
+    _call_remove_across,
+    _cache_get,
+    _cache_put,
+    _del_trakt,
+    _headers,
+    _is_debug,
+    _load_config,
+    _save_config,
+    _tokens,
+)
+
 TRAKT_API = "https://api.trakt.tv"
 
 _SCROBBLE_STATE: dict[str, dict[str, Any]] = {}
@@ -61,61 +75,6 @@ _DEF_TRAKT: dict[str, Any] = {
     "force_stop_at": 95,
     "regress_tolerance_percent": 5,
 }
-
-
-def _call_remove_across(ids: dict[str, Any], media_type: str) -> None:
-    if not isinstance(ids, dict) or not ids:
-        return
-    try:
-        cfg = _load_config()
-        s = (cfg.get("scrobble") or {})
-        if not s.get("delete_plex"):
-            return
-        tps = s.get("delete_plex_types") or []
-        mt = (media_type or "").strip().lower()
-        allow = False
-        if isinstance(tps, list):
-            allow = (mt in tps) or ((mt.rstrip("s") + "s") in tps)
-        elif isinstance(tps, str):
-            allow = mt in tps
-        if not allow:
-            return
-    except Exception:
-        pass
-    try:
-        if callable(_rm_across):
-            _rm_across(ids, media_type)
-            return
-    except Exception:
-        pass
-    try:
-        if callable(_rm_across_api):
-            _rm_across_api(ids, media_type)  # type: ignore[arg-type]
-            return
-    except Exception:
-        pass
-
-
-def _load_config() -> dict[str, Any]:
-    try:
-        return load_config()
-    except Exception:
-        return {}
-
-
-def _save_config(cfg: dict[str, Any]) -> None:
-    try:
-        save_config(cfg)
-    except Exception:
-        pass
-
-
-def _is_debug() -> bool:
-    try:
-        rt = (_load_config().get("runtime") or {})
-        return bool(rt.get("debug") or rt.get("debug_mods"))
-    except Exception:
-        return False
 
 
 def _emit(logger: Callable[..., None] | Any | None, msg: str, level: str = "INFO") -> None:
@@ -226,59 +185,6 @@ def _as_filter_set(value: Any) -> set[str]:
             if clean:
                 out.add(clean)
     return out
-
-
-def _tokens(cfg: dict[str, Any]) -> dict[str, str]:
-    tr = cfg.get("trakt") or {}
-    au = ((cfg.get("auth") or {}).get("trakt") or {})
-    return {
-        "client_id": (tr.get("client_id") or "").strip(),
-        "client_secret": (tr.get("client_secret") or "").strip(),
-        "access_token": (au.get("access_token") or tr.get("access_token") or "").strip(),
-        "refresh_token": (au.get("refresh_token") or tr.get("refresh_token") or "").strip(),
-    }
-
-
-def _app_meta(cfg: dict[str, Any]) -> dict[str, str]:
-    rt = (cfg.get("runtime") or {})
-    av = str(rt.get("version") or "CrossWatch/Scrobble")
-    ad = (rt.get("build_date") or "").strip()
-    meta: dict[str, str] = {"app_version": av}
-    if ad:
-        meta["app_date"] = ad
-    return meta
-
-
-def _headers(cfg: dict[str, Any]) -> dict[str, str]:
-    t = _tokens(cfg)
-    h = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "trakt-api-version": "2",
-        "trakt-api-key": t["client_id"],
-        "User-Agent": "CrossWatch/Scrobble",
-    }
-    if t["access_token"]:
-        h["Authorization"] = f"Bearer {t['access_token']}"
-    return h
-
-
-def _del_trakt(path: str, cfg: dict[str, Any]) -> requests.Response:
-    url = f"{TRAKT_API}{path}"
-    r = requests.delete(url, headers=_headers(cfg), timeout=12)
-    if r.status_code == 401:
-        try:
-            from providers.auth._auth_TRAKT import PROVIDER as TRAKT_AUTH
-
-            TRAKT_AUTH.refresh(cfg)
-            _save_config(cfg)
-        except Exception:
-            return r
-        try:
-            r = requests.delete(url, headers=_headers(cfg), timeout=12)
-        except Exception:
-            pass
-    return r
 
 
 def _get_trakt_watching(cfg: dict[str, Any]) -> None:
@@ -631,20 +537,6 @@ def _probe_played_status(cfg: dict[str, Any], rating_key: Any) -> bool:
         return False
 
 
-def _cache_get(key: tuple[Any, ...]) -> Any | None:
-    try:
-        return _TRAKT_ID_CACHE.get(key)
-    except Exception:
-        return None
-
-
-def _cache_put(key: tuple[Any, ...], value: Any) -> None:
-    try:
-        if len(_TRAKT_ID_CACHE) > 2048:
-            _TRAKT_ID_CACHE.clear()
-        _TRAKT_ID_CACHE[key] = value
-    except Exception:
-        pass
 
 
 def _resolve_trakt_movie_id(
@@ -653,7 +545,7 @@ def _resolve_trakt_movie_id(
     logger: Callable[..., None] | Any | None = None,
 ) -> int | None:
     key = ("movie", ids_all.get("imdb"), ids_all.get("tmdb"), ids_all.get("tvdb"))
-    c = _cache_get(key)
+    c = _cache_get(_TRAKT_ID_CACHE, key)
     if c is not None:
         return c
     for k in ("tmdb", "imdb", "tvdb"):
@@ -674,11 +566,11 @@ def _resolve_trakt_movie_id(
                 continue
             tid = (((arr[0] or {}).get("movie") or {}).get("ids") or {}).get("trakt")
             if tid:
-                _cache_put(key, int(tid))
+                _cache_put(_TRAKT_ID_CACHE, key, int(tid))
                 return int(tid)
         except Exception as e:
             _emit(logger, f"trakt movie id resolve error: {e}", "DEBUG")
-    _cache_put(key, None)
+    _cache_put(_TRAKT_ID_CACHE, key, None)
     return None
 
 
@@ -688,7 +580,7 @@ def _resolve_trakt_show_id(
     logger: Callable[..., None] | Any | None = None,
 ) -> int | None:
     key = ("show", ids_all.get("imdb"), ids_all.get("tmdb"), ids_all.get("tvdb"))
-    c = _cache_get(key)
+    c = _cache_get(_TRAKT_ID_CACHE, key)
     if c is not None:
         return c
     for k in ("tmdb", "imdb", "tvdb"):
@@ -709,11 +601,11 @@ def _resolve_trakt_show_id(
                 continue
             tid = (((arr[0] or {}).get("show") or {}).get("ids") or {}).get("trakt")
             if tid:
-                _cache_put(key, int(tid))
+                _cache_put(_TRAKT_ID_CACHE, key, int(tid))
                 return int(tid)
         except Exception as e:
             _emit(logger, f"trakt show id resolve error: {e}", "DEBUG")
-    _cache_put(key, None)
+    _cache_put(_TRAKT_ID_CACHE, key, None)
     return None
 
 
@@ -727,7 +619,7 @@ def _trakt_show_ids_from_imdb_show(
         return {}
 
     key = ("show_ids_imdb", imdb_show)
-    c = _cache_get(key)
+    c = _cache_get(_TRAKT_ID_CACHE, key)
     if isinstance(c, dict):
         return c
     if c is not None:
@@ -741,23 +633,23 @@ def _trakt_show_ids_from_imdb_show(
             timeout=10,
         )
         if r.status_code != 200:
-            _cache_put(key, None)
+            _cache_put(_TRAKT_ID_CACHE, key, None)
             return {}
         arr = r.json() or []
         if not arr:
-            _cache_put(key, None)
+            _cache_put(_TRAKT_ID_CACHE, key, None)
             return {}
 
         ids = (((arr[0] or {}).get("show") or {}).get("ids") or {})
         out = {k: ids[k] for k in ("trakt", "tmdb", "imdb", "tvdb") if ids.get(k)}
 
-        _cache_put(key, out if out else None)
+        _cache_put(_TRAKT_ID_CACHE, key, out if out else None)
         if out:
             _emit(logger, f"trakt show ids from imdb_show {imdb_show}: {out}", "DEBUG")
         return out
     except Exception as e:
         _emit(logger, f"trakt show ids from imdb_show {imdb_show} error: {e}", "DEBUG")
-        _cache_put(key, None)
+        _cache_put(_TRAKT_ID_CACHE, key, None)
         return {}
 
 
@@ -866,14 +758,14 @@ def _resolve_trakt_episode_id(
     except Exception:
         e = None
     key = ("episode", ids_all.get("imdb"), ids_all.get("tmdb"), ids_all.get("tvdb"), s, e)
-    c = _cache_get(key)
+    c = _cache_get(_TRAKT_ID_CACHE, key)
     if c is not None:
         return c
     hint = {**_episode_ids_from_md(md), **ids_all, **_all_ids_from_metadata(md)}
     found = _guid_search_episode(hint, cfg, logger=logger)
     tid = (found or {}).get("trakt")
     if isinstance(tid, int):
-        _cache_put(key, tid)
+        _cache_put(_TRAKT_ID_CACHE, key, tid)
         return tid
     show_tid = _resolve_trakt_show_id(ids_all, cfg, logger=logger)
     if show_tid and isinstance(s, int) and isinstance(e, int):
@@ -887,16 +779,12 @@ def _resolve_trakt_episode_id(
                 ej = r.json() or {}
                 tid2 = ((ej.get("ids") or {}).get("trakt"))
                 if tid2:
-                    _cache_put(key, int(tid2))
+                    _cache_put(_TRAKT_ID_CACHE, key, int(tid2))
                     return int(tid2)
         except Exception as ex:
             _emit(logger, f"trakt ep id resolve error: {ex}", "DEBUG")
-    _cache_put(key, None)
+    _cache_put(_TRAKT_ID_CACHE, key, None)
     return None
-
-
-def _best_id_key_order(media_type: str) -> tuple[str, ...]:
-    return ("tmdb", "imdb", "tvdb") if media_type == "movie" else ("tmdb", "imdb", "tvdb")
 
 
 def _build_primary_body(
